feat(#4250): Refactor backend and frontend code
