+++ streampipes-rest/src/main/java/org/apache/streampipes/rest/impl/datalake/importer/CsvImportValidationException.java
@@ -1,0 +1,37 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package org.apache.streampipes.rest.impl.datalake.importer;
+
+import org.apache.streampipes.model.datalake.importer.CsvImportValidationMessage;
+
+import java.util.List;
+
+public class CsvImportValidationException extends RuntimeException {
+
+  private final List<CsvImportValidationMessage> validationMessages;
+
+  public CsvImportValidationException(List<CsvImportValidationMessage> validationMessages) {
+    super(validationMessages.isEmpty() ? "CSV import validation failed" : validationMessages.get(0).getMessage());
+    this.validationMessages = validationMessages;
+  }
+
+  public List<CsvImportValidationMessage> getValidationMessages() {
+    return validationMessages;
+  }
+}
