+++ ui/src/app/dataset/dialog/csv-import-dialog/csv-import-preview-table/csv-import-preview-table.component.scss
@@ -1,0 +1,77 @@
+/*!
+ *  Licensed to the Apache Software Foundation (ASF) under one or more
+ *  contributor license agreements.  See the NOTICE file distributed with
+ *  this work for additional information regarding copyright ownership.
+ *  The ASF licenses this file to You under the Apache License, Version 2.0
+ *  (the "License"); you may not use this file except in compliance with
+ *  the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
+
+.schema-mismatch-list {
+    margin: 10px 0 0;
+    padding-left: 20px;
+    color: var(--color-text-200, #5a6673);
+}
+
+.schema-mismatch-item + .schema-mismatch-item {
+    margin-top: 6px;
+}
+
+.preview-table-wrapper {
+    overflow: auto;
+    border: 1px solid var(--color-border, #d8dee4);
+    border-radius: 14px;
+}
+
+.preview-table {
+    width: max-content;
+    min-width: 100%;
+    border-collapse: collapse;
+}
+
+.preview-table th,
+.preview-table td {
+    padding: 10px 12px;
+    border-bottom: 1px solid var(--color-border, #d8dee4);
+    text-align: left;
+    vertical-align: top;
+    white-space: nowrap;
+    font-size: 13px;
+}
+
+.preview-table th {
+    position: sticky;
+    top: 0;
+    background: #f8fafc;
+    z-index: 1;
+}
+
+.header-cell {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    min-width: 220px;
+}
+
+.header-title {
+    font-weight: 700;
+    font-size: 14px;
+    color: #111827;
+}
+
+.empty-state {
+    padding: 20px;
+    border-radius: 12px;
+    background: #f8fafc;
+    color: var(--color-text-200, #5a6673);
+    font-size: 14px;
+}
