+++ ui/src/app/dataset/dialog/csv-import-dialog/csv-import-upload-state/csv-import-upload-state.component.html
@@ -1,0 +1,56 @@
+<!--
+  ~ Licensed to the Apache Software Foundation (ASF) under one or more
+  ~ contributor license agreements.  See the NOTICE file distributed with
+  ~ this work for additional information regarding copyright ownership.
+  ~ The ASF licenses this file to You under the Apache License, Version 2.0
+  ~ (the "License"); you may not use this file except in compliance with
+  ~ the License.  You may obtain a copy of the License at
+  ~
+  ~    http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  -->
+
+<div class="upload-state">
+    @if (importLoading()) {
+        <mat-spinner [diameter]="56"></mat-spinner>
+        <div class="upload-title">
+            {{ 'Uploading CSV data' | translate }}
+        </div>
+        <div class="upload-copy">
+            {{
+                'The data is currently being written to the data lake.'
+                    | translate
+            }}
+        </div>
+    } @else if (hasImportResult()) {
+        <mat-icon class="success-icon">check_circle</mat-icon>
+        <div class="upload-title" data-cy="csv-import-success-title">
+            {{ 'Import completed' | translate }}
+        </div>
+        <div class="result-name">
+            {{ importResult()?.measurementName }}
+        </div>
+        <div class="result-meta">
+            {{ importResult()?.importedRowCount }}
+            {{ 'rows imported' | translate }}
+        </div>
+    } @else if (uploadErrors().length > 0) {
+        <sp-alert-banner
+            type="error"
+            [title]="'Error' | translate"
+            [description]="'The CSV import failed.' | translate"
+            data-cy="csv-import-upload-error"
+        ></sp-alert-banner>
+        <div class="upload-errors" data-cy="csv-import-upload-error-details">
+            @for (message of uploadErrors(); track message.message) {
+                <div class="upload-error-line">{{ message.message }}</div>
+            }
+        </div>
+    }
+</div>
